app/requirements.txt
django-redis==5.3.0
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
celery==5.3.0
certifi==2023.7.22
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.4
django-jazzmin==2.6.0
docx2txt==0.8
frozenlist==1.4.0
idna==3.4
kombu==5.3.2
lxml==4.9.3
multidict==6.0.4
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
PyPDF2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
redis==5.0.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
tzdata==2023.3
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
